requirements.txt
adafruit-pca9685
rpi_ws281x
mpu6050-raspberrypi

flask
flask_cors
websockets

gtts
pygame

numpy
opencv-contrib-python==3.4.3.18
